pyproject.toml
[tool.poetry]
name = "dotdict-bench"
version = "0.0.0"
description = "Benchmarking for dot-accessible dict packages in python"
authors = ["pwwang <pwwang@pwwang.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dotdict_bench"}]

[tool.poetry.dependencies]
python = "^3.9"
benchwork = "^0.0.1"
addict = "^2.4.0"
python-box = "^6.0.2"
glom = "^22.1.0"
dotmap = "^1.3.30"
dotwiz = "^0.4.0"
dotsi = "^0.0.3"
dictlib = "^1.1.5"
diot = "^0.1.6"
importlib-metadata = "^4.12.0"
PyGithub = "^1.55"
pyparam = "^0.5.3"
easydict = "^1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
